projects/3d-civil-works/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
pillow==10.1.0
numpy>=1.26.0
pytest==7.4.3
httpx==0.25.0
